requirements.txt
appdirs==1.4.4
asgiref==3.4.1
attrs==21.2.0
autoflake==1.4
autopep8==1.5.7
black==21.7b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.5
django-stubs==1.8.0
django-stubs-ext==0.2.0
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.11.0
flake8==3.9.2
idna==3.2
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
pluggy==0.13.1
pprintpp==0.4.0
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.9.0
pyparsing==2.4.7
pytest==6.2.4
pytest-clarity==1.0.1
pytest-django==4.4.0
pytest-factoryboy==2.1.0
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.3
requests==2.26.0
rich==10.7.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
yapf==0.31.0
